+++ app/src/main/java/me/yifeiyuan/flapdev/components/simpletext/SimpleTextComponent.java
@@ -1,19 +1,20 @@
 package me.yifeiyuan.flapdev.components.simpletext;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
 
-import me.yifeiyuan.flap.FlapComponent;
-import me.yifeiyuan.flap.annotations.Component;
+import me.yifeiyuan.flap.Component;
+import me.yifeiyuan.flap.annotations.Proxy;
 import me.yifeiyuan.flapdev.R;
 
 /**
  * Created by 程序亦非猿 on 2018/12/4.
  */
-@Component(layoutId = R.layout.flap_item_simple_text, autoRegister = true)
-public class SimpleTextComponent extends FlapComponent<SimpleTextModel> {
+//@Proxy(layoutId = R.layout.flap_item_simple_text,layoutName = "flap_item_simple_text")
+@Proxy(layoutName = "flap_item_simple_text")
+public class SimpleTextComponent extends Component<SimpleTextModel> {
 
     private static final String TAG = "SimpleTextItem";