+++ flap-compiler/src/main/java/me/yifeiyuan/flap/compiler/FlapProcessor.java
@@ -9,7 +9,6 @@ import com.squareup.javapoet.TypeName;
 import com.squareup.javapoet.TypeSpec;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
@@ -32,45 +31,40 @@ import javax.lang.model.util.SimpleTypeVisitor8;
 import javax.lang.model.util.Types;
 import javax.tools.Diagnostic;
 
-import me.yifeiyuan.flap.annotations.AutoRegister;
-import me.yifeiyuan.flap.annotations.Component;
+import me.yifeiyuan.flap.annotations.Proxy;
 
 @AutoService(Processor.class)
 public class FlapProcessor extends AbstractProcessor {
 
     private static final String PKG_NAME_PROXIES = "me.yifeiyuan.flap.apt.proxies";
-    private static final String PKG_NAME_MANAGER = "me.yifeiyuan.flap.apt.manager";
 
     private static final String NAME_SUFFIX = "Proxy";
 
-    private final ClassName CLASS_KEEP = ClassName.bestGuess("android.support.annotation.Keep");
+    private final ClassName CLASS_KEEP = ClassName.bestGuess("androidx.annotation.Keep");
     private final ClassName CLASS_FLAP = ClassName.bestGuess("me.yifeiyuan.flap.Flap");
     private final ClassName CLASS_COMPONENT_PROXY = ClassName.bestGuess("me.yifeiyuan.flap.internal.ComponentProxy");
 
-    private static final String KEY_OPTION_AUTO_REGISTER = "autoRegister";
+    private static final String KEY_OPTION_R_CLASS_PATH = "packageName";
 
     private Filer filer;
-    private Elements elementUtils;
-    private Types typeUtils;
+    private Elements elements;
+    private Types types;
     private Messager messager;
 
-    /**
-     * 是否自动注册 Factories
-     */
-    private boolean autoRegisterFactories = true;
+    private String rPackageName;
 
     @Override
     public synchronized void init(final ProcessingEnvironment processingEnv) {
         super.init(processingEnv);
-        filer = processingEnv.getFiler();               // Generate class.
-        typeUtils = processingEnv.getTypeUtils();            // Get type utils.
-        elementUtils = processingEnv.getElementUtils();      // Get class meta.
+        filer = processingEnv.getFiler();
+        types = processingEnv.getTypeUtils();
+        elements = processingEnv.getElementUtils();
         messager = processingEnv.getMessager();
         messager.printMessage(Diagnostic.Kind.NOTE, "FlapProcessor init");
 
         Map<String, String> options = processingEnv.getOptions();
-        if (options.containsKey(KEY_OPTION_AUTO_REGISTER)) {
-            autoRegisterFactories = Boolean.parseBoolean(options.get(KEY_OPTION_AUTO_REGISTER));
+        if (options.containsKey(KEY_OPTION_R_CLASS_PATH)) {
+            rPackageName = options.get(KEY_OPTION_R_CLASS_PATH);
         }
     }
 
@@ -79,11 +73,10 @@ public class FlapProcessor extends AbstractProcessor {
 
         for (final TypeElement typeElement : set) {
 
-            if (Component.class.getCanonicalName().equals(typeElement.getQualifiedName().toString())) {
+            if (Proxy.class.getCanonicalName().equals(typeElement.getQualifiedName().toString())) {
                 processComponent(roundEnvironment, typeElement);
-            } else if (AutoRegister.class.getCanonicalName().equals(typeElement.getQualifiedName().toString())) {
-                processComponentProxyManager(roundEnvironment, typeElement);
             }
+
         }
 
         return true;
@@ -91,14 +84,15 @@ public class FlapProcessor extends AbstractProcessor {
 
     private void processComponent(final RoundEnvironment roundEnvironment, final TypeElement typeElement) {
 
-        Set<? extends Element> elements = roundEnvironment.getElementsAnnotatedWith(Component.class);
+        Set<? extends Element> elements = roundEnvironment.getElementsAnnotatedWith(Proxy.class);
 
         for (final Element element : elements) {
-            Component factory = element.getAnnotation(Component.class);
+            Proxy factory = element.getAnnotation(Proxy.class);
             if (null != factory) {
                 try {
                     TypeSpec flapItemFactoryTypeSpec = createComponentProxyTypeSpec(roundEnvironment, typeElement, (TypeElement) element, factory);
                     JavaFile.builder(PKG_NAME_PROXIES, flapItemFactoryTypeSpec)
+                            .addFileComment("由 Flap APT 自动生成，请勿修改。")
                             .build()
                             .writeTo(filer);
                 } catch (IOException e) {
@@ -115,10 +109,9 @@ public class FlapProcessor extends AbstractProcessor {
      * @param typeElement          @Component
      * @param flapComponentElement 被 FlapComponent 注解了的那个类
      * @param componentProxy       注解了目标类的 注解，可以获取值
-     *
      * @return ComponentProxy TypeSpec
      */
-    private TypeSpec createComponentProxyTypeSpec(final RoundEnvironment roundEnvironment, final TypeElement typeElement, final TypeElement flapComponentElement, final Component componentProxy) {
+    private TypeSpec createComponentProxyTypeSpec(final RoundEnvironment roundEnvironment, final TypeElement typeElement, final TypeElement flapComponentElement, final Proxy componentProxy) {
 
         ClassName flapItemClass = (ClassName) ClassName.get(flapComponentElement.asType());
 
@@ -126,13 +119,14 @@ public class FlapProcessor extends AbstractProcessor {
         String targetClassName = flapComponentElement.getSimpleName().toString() + NAME_SUFFIX;
 
         int layoutId = componentProxy.layoutId();
-        boolean autoRegister = componentProxy.autoRegister();
 
-        boolean dataBinding = componentProxy.useDataBinding();
+        String layoutName = componentProxy.layoutName();
+
+        boolean useDataBinding = componentProxy.useDataBinding();
 
         DeclaredType declaredType = flapComponentElement.getSuperclass().accept(new FlapItemModelVisitor(), null);
         List<? extends TypeMirror> args = declaredType.getTypeArguments();
-        TypeElement itemModelType = (TypeElement) typeUtils.asElement(args.get(0));
+        TypeElement itemModelType = (TypeElement) types.asElement(args.get(0));
 
         ClassName itemModelClass = ClassName.get(itemModelType);
 
@@ -147,21 +141,35 @@ public class FlapProcessor extends AbstractProcessor {
                 .addParameter(TypeName.INT, "layoutId")
                 .returns(flapItemClass);
 
-        if (dataBinding) {
-            onCreateViewHolderMethodBuilder.addStatement("return new $T(android.databinding.DataBindingUtil.inflate(inflater,layoutId,parent,false))",flapItemClass);
+        if (useDataBinding) {
+            onCreateViewHolderMethodBuilder.addStatement("return new $T(androidx.databinding.DataBindingUtil.inflate(inflater,layoutId,parent,false))", flapItemClass);
         } else {
-            onCreateViewHolderMethodBuilder.addStatement("return new $T(inflater.inflate(layoutId,parent,false))", flapItemClass);
+
+            boolean useViewBinding = componentProxy.useViewBinding();
+
+            if (useDataBinding) {
+                // TODO: 2020/9/27
+
+            } else {
+                onCreateViewHolderMethodBuilder.addStatement("return new $T(inflater.inflate(layoutId,parent,false))", flapItemClass);
+            }
         }
 
         MethodSpec onCreateViewHolderMethod = onCreateViewHolderMethodBuilder.build();
 
-        MethodSpec getItemViewTypeMethod = MethodSpec.methodBuilder("getItemViewType")
+        MethodSpec.Builder getItemViewTypeMethodBuilder = MethodSpec.methodBuilder("getItemViewType")
                 .addAnnotation(Override.class)
                 .addModifiers(Modifier.PUBLIC)
                 .addParameter(itemModelClass, "model")
-                .returns(Integer.TYPE)
-                .addStatement("return " + layoutId)
-                .build();
+                .returns(Integer.TYPE);
+
+        if ("".equals(layoutName)) {
+            getItemViewTypeMethodBuilder.addStatement("return " + layoutId);
+        } else {
+            getItemViewTypeMethodBuilder.addStatement("return " + rPackageName + ".R.layout." + layoutName);
+        }
+
+        MethodSpec getItemViewTypeMethod = getItemViewTypeMethodBuilder.build();
 
         MethodSpec getComponentModelClass = MethodSpec.methodBuilder("getComponentModelClass")
                 .addAnnotation(Override.class)
@@ -181,58 +189,6 @@ public class FlapProcessor extends AbstractProcessor {
                         .addMethod(getComponentModelClass)
                         .addSuperinterface(name);
 
-        if (autoRegister) {
-            builder.addAnnotation(AutoRegister.class);
-        }
-        return builder.build();
-    }
-
-    /**
-     * 处理 AutoRegister 注解，把需要自动注册的组件处理一下。
-     *
-     * @param roundEnvironment
-     * @param typeElement
-     */
-    private void processComponentProxyManager(final RoundEnvironment roundEnvironment, final TypeElement typeElement) {
-
-        if (!autoRegisterFactories) {
-            return;
-        }
-        List<ClassName> factories = new ArrayList<>();
-
-        Set<? extends Element> elements = roundEnvironment.getElementsAnnotatedWith(AutoRegister.class);
-
-        for (final Element element : elements) {
-            TypeElement flapItemFactory = (TypeElement) element;
-            ClassName factoryClass = ClassName.get(flapItemFactory);
-            factories.add(factoryClass);
-        }
-
-        TypeSpec manager = TypeSpec.classBuilder("ComponentAutoRegister")
-                .addModifiers(Modifier.PUBLIC, Modifier.FINAL)
-                .addAnnotation(CLASS_KEEP)
-                .addMethod(createInjectMethod(factories))
-                .build();
-
-        try {
-            JavaFile.builder(PKG_NAME_MANAGER, manager).build().writeTo(filer);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-    }
-
-    private MethodSpec createInjectMethod(final List<ClassName> factories) {
-
-        MethodSpec.Builder builder = MethodSpec.methodBuilder("inject")
-                .addModifiers(Modifier.PUBLIC, Modifier.STATIC, Modifier.FINAL)
-                .returns(void.class)
-                .addParameter(CLASS_FLAP, "flap", Modifier.FINAL);
-
-        for (final ClassName factory : factories) {
-            builder.addStatement("flap.register(new $T())", factory);
-        }
-
         return builder.build();
     }
 
@@ -244,8 +200,7 @@ public class FlapProcessor extends AbstractProcessor {
     @Override
     public Set<String> getSupportedAnnotationTypes() {
         Set<String> annotationTypes = new LinkedHashSet<>();
-        annotationTypes.add(Component.class.getCanonicalName());
-        annotationTypes.add(AutoRegister.class.getCanonicalName());
+        annotationTypes.add(Proxy.class.getCanonicalName());
         return annotationTypes;
     }