+++ othermodule/src/main/java/me/yifeiyuan/ktx/foundation/othermodule/JavaModuleModel.java
@@ -1,0 +1,7 @@
+package me.yifeiyuan.ktx.foundation.othermodule;
+
+/**
+ * Created by 程序亦非猿 on 2020/9/21.
+ */
+public class JavaModuleModel {
+}